Added Update Location button
The file generated by the layout program might not locate a part
correctly.  This button allows one to jog to the right location
and click it.  It takes the difference between the measured location
and the current CNC location and applies it to the nominal location.
The current CNC location becomes the new measured location.  Any
future application of the fiducial measurement algorithm to the
new nominal location should result in the proper measured location.