Slack Compensation in all directions
The Slack Compensation algorithm didn't apply in the positive X and Y
direction (1st quadrant). However it was found that the positioning was
more accurate with the small move at the end due to slack compensation.
This change makes that small move for all moves.  This includes moves
that are in increasing X and Y directions.