Adjusting Parts over Up Camera
This commit is for a simple change to allow an adjusted offset based
on the part's position over the up camera.  This adjustment is for
manual placement only and only works with the Place Here button. The
variable that indicates the adjusted state is cleared on any change
in the vacuum state. A change in the vacuum state indicates that the
present part has been placed or a new part picked up.  Either way
the adjustment is no longer valid.