Add ability to rotate camera cross
Added the ability to rotate the cross in the camera displays.  Added
a checkbox to allow jogging keys to control the rotation.  There is
a flag in the camera class that allows the cross to be under programmatic
control.  This would apply to the up camera for adjusting parts for
placement.  The cross would rotate to the angle needed for the given
part